Finish 1st working version of the benchmark

+++ leniax/runner.py
@@ -62,7 +62,7 @@ def run(
 
     previous_mass = init_mass
     previous_sign = jnp.zeros(N, dtype=jnp.int32)
-    counters = leniax_stat.init_counters(N, dtype=jnp.int32)
+    counters = leniax_stat.init_counters(N)
     should_continue = jnp.ones(N, dtype=jnp.int32)
     total_shift_idx = jnp.zeros([N, nb_world_dims], dtype=jnp.int32)
     mass_centroid = jnp.zeros([nb_world_dims, N], dtype=jnp.int32)
